prep v0.5.0 release: bump Plonky3 deps

+++ p3-miden-dev-utils/src/fixtures.rs
@@ -3,8 +3,7 @@
 //! This module contains constants and helper functions that define
 //! reproducible test/benchmark scenarios.
 
-use alloc::vec;
-use alloc::vec::Vec;
+use alloc::{vec, vec::Vec};
 
 use p3_field::PackedValue;
 
@@ -23,7 +22,7 @@ pub const BENCH_SEED: u64 = TEST_SEED;
 // =============================================================================
 
 /// Standard log heights for benchmarking: 2^16, 2^18, 2^20 leaves.
-pub const LOG_HEIGHTS: &[usize] = &[16, 18, 20];
+pub const LOG_HEIGHTS: &[u8] = &[16, 18, 20];
 
 /// Standard relative specs for benchmark matrix groups.
 ///
